cfehome/requirements.txt
boto3==1.9.42
botocore==1.12.42
Django==2.0.9
docutils==0.14
jmespath==0.9.3
numpy==1.15.4
picamera==1.13
Pillow==5.3.0
python-dateutil==2.7.5
pytz==2018.7
s3transfer==0.1.13
six==1.11.0
urllib3==1.24.1
